pyproject.toml
[project]
name = "google-adk-yfinance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.14.1",
    "yfinance>=0.2.66",
]
